==> R/05_analyze_richness_2x7x3.Rmd <==
---
title: "Richness by combined zones"
abstract: |
  This notebook uses species richness layers created by SDM and polygons for three landscape variables to compare species richness distributions within the zones created by the three variables.  
output: 
  html_document:
    code_folding: hide
    toc: true
date: "`r format(Sys.time(), '%d %B %Y')`"
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      fig.align = 'center',
                      fig.height=8, 
                      fig.width=7, 
                      message=FALSE, 
                      warning=FALSE, 
                      cache=TRUE)

```

This notebook uses species richness layers created by SDM and polygons for three landscape variables to compare species richness distributions within the zones created by the three variables.

The landscape variables/polygons are natural protected areas (ANPs), the 7 ecoregiones (aka. biomes) of Mexico, and landcover classified into vegetation, agriculture, and other. The ANP variable considers inside protected areas, within 10 km of the areas, and greater than 10 km from a protected area. These zone polygons were previously created (prep_create_zone_polygons.R).

The richness layers were created by stacking species distribution models (SDMs) that were produced for every species considered (process_SDM_v2.R).

The result of the notebook is a collection of graphics displaying richness for each zone and indicating whether the differences in richness are statistically significant.

```{r initialize, message=FALSE}
# Load libraries and initialize variables
source('R/00_initialize.R')

samp_n <- 100
# samp_n <- NA
# sessionInfo()

caption <- "Different colors within one facet indicate statistically significant (p < 0.05) difference between groups."

bioms_english_sorted <- c(
      'Mediterranean California',
      'Deserts',
      'Great Plains',
      'Semi-arid Highlands',
      'Temperate Sierras',
      'Tropical Dry Forests',
      'Tropical Wet Forests'
    )

# # Zones raster and lookup table
# zone_polys_dir <- file.path('data', 'tidy', 'analysis_zones', 'from_v1')
# zones_fn <- 'v1_stack_anps3_biomes7_usv3'
# zones_ras_fp <- file.path(zone_polys_dir, str_c(zones_fn, '.tif'))
# zones_lu_fp <- file.path(zone_polys_dir, str_c(zones_fn, '_lu.rds'))
```

Parameters:

-   unq_cells: **`r unq_cells`**
-   excludep: **`r excludep`**
-   filt_dates: **`r filt_dates`**
-   buffer_distance: **`r buffer_distance` km**

```{r functions}
extract_stats_to_csv <- function(rich_tif_fp, div_dir, combo_ras, 
                                 rich_dir = NA, 
                                 zones_code = '',
                                 col_types = 'ifffdf', overwrite = FALSE){
  
  # Resulting path for CSV
  rich_fn <- basename(tools::file_path_sans_ext(rich_tif_fp))
  if(is.na(rich_dir))  rich_dir <- dirname(rich_tif_fp)
  div_df_fp <- file.path(rich_dir, paste0('csvs_', zones_code), str_c(rich_fn, '.csv'))
  dir.create(dirname(div_df_fp), recursive = TRUE, showWarnings = FALSE)

  if(file.exists(div_df_fp) & !overwrite) return(read_csv(div_df_fp, col_types=col_types))
  
  # Load raster and add to SpatRaster
  rich_ras <- terra::rast(rich_tif_fp)
  names(rich_ras) <- 'richness'
  combo_all <- c(combo_ras, rich_ras)
  
  # Convert to DF and recode
  # pol_group <- str_extract(rich_fn, 'bee|butterfly|moth|hummingbird|bat|wasp|fly')
  pol_group <- str_extract(rich_fn, '(?<=richness_).*(?=_rf)')
  div_df <- terra::as.data.frame(combo_all, cells=TRUE) %>% 
    mutate(pol_group = factor(pol_group))
  
  # Save
  div_df %>% write_csv(div_df_fp)

  return(div_df) 
}

create_richness_df <- function(rich_dir, div_dir, zones_lu_fp, combo_ras, df_fp = NULL, 
                               zones_code = '', overwrite = FALSE) {
  
  if(is.null(df_fp))  df_fp <- file.path(rich_dir, paste0('richness_df', zones_code, '.rds'))
  if(file.exists(df_fp) & !overwrite) return(readRDS(df_fp))
  
  # List richness tifs (1 per taxonomic group)
  (fps <- list.files(rich_dir, '*\\.tif', full.names = TRUE))
  
  # Create/load DFs
  div_df <- fps %>% purrr::map_dfr(extract_stats_to_csv, div_dir, combo_ras, 
                                   zones_code = zones_code, overwrite = FALSE)
  
  # Recode biome names
  lu <- readRDS(zones_lu_fp)
  biom_codes_all <- tibble(
    id = names(lu$biom),
    name = lu$biom,
    code = c(
      'CalMedit',
      'Desierto',
      'ElevSemiar',
      'GranPlanic',
      'SelvCalHum',
      'SelvCalSec',
      'SierTemp'
    ),
    english = c(
      'Mediterranean California',
      'Deserts',
      'Semi-arid Highlands',
      'Great Plains',
      'Tropical Wet Forests',
      'Tropical Dry Forests',
      'Temperate Sierras'
    )
  )
  biom_codes <- setNames(biom_codes_all$english, biom_codes_all$id)
  
  # Convert numeric code to factors
  div_df <- div_df %>%
    mutate(
      anp_zone = recode(anp_zone, !!!lu$anp) %>%
        factor(levels = c(
          # 'Outside buffer', 'Buffer 10 km', 'Inside NPA'
          'Not NPA', 'NPA'
        )),
      biome = recode(biome, !!!biom_codes) %>%
        factor(),
      landcover = recode(landcover, !!!lu$landcover) %>%
        factor(levels = c('veg', 'ag', 'otro'))
    )

  # Save
  div_df %>% saveRDS(df_fp)
  
  # Return
  return(div_df)
}


create_richness_df_with_ids <- function(rich_dir, div_dir, zones_lu_fp, combo_ras, 
                                        df_fp = NULL,
                                        zones_code = '', overwrite = FALSE) {
  
  if(is.null(df_fp))  df_fp <- file.path(rich_dir, paste0('richness_df', zones_code, '.rds'))
  if(file.exists(df_fp) & !overwrite) return(readRDS(df_fp))
  
  # List richness tifs
  (fps <- list.files(rich_dir, '*\\.tif', full.names = TRUE))
  
  # Create/load DFs
  div_df <- fps %>% purrr::map_dfr(extract_stats_to_csv, div_dir, combo_ras,
                                   rich_dir = div_dir, col_type = 'iffffdf', 
                                   zones_code = zones_code, overwrite= FALSE)
  
  # Recode biome names
  lu <- readRDS(zones_lu_fp)
  biom_codes_all <- tibble(
    id = names(lu$biom),
    name = lu$biom,
    code = c(
      'CalMedit',
      'Desierto',
      'ElevSemiar',
      'GranPlanic',
      'SelvCalHum',
      'SelvCalSec',
      'SierTemp'
    ),
    english = c(
      'Mediterranean California',
      'Deserts',
      'Semi-arid Highlands',
      'Great Plains',
      'Tropical Wet Forests',
      'Tropical Dry Forests',
      'Temperate Sierras'
    )
  )
  biom_codes <- setNames(biom_codes_all$english, biom_codes_all$id)
  
  # Convert numeric code to factors
  div_df <- div_df %>%
    mutate(
      anp_zone = recode(anp_zone, !!!lu$anp) %>%
        factor(levels = c(
          # 'Outside buffer', 'Buffer 10 km', 'Inside NPA'
          'Not NPA', 'NPA'
        )),
      biome = recode(biome, !!!biom_codes) %>%
        factor(),
      landcover = recode(landcover, !!!lu$landcover) %>%
        factor(levels = c('veg', 'ag', 'otro')),
      ID_ANP = recode(ID_ANP, !!!lu$ID_ANP)
    )
  
  # Save
  div_df %>% saveRDS(df_fp)
  
  # Return
  return(div_df)
}
  
sample_df <- function(div_df, samp_n, samp_fp, 
                      grp_vars = c('anp_zone', 'biome', 'landcover'),
                      overwrite = FALSE) {
  # Sample the same cells for all the pollinator groups
  if(file.exists(samp_fp) & !overwrite) return(readRDS(samp_fp))
  
  # Get all usable cells
  polx <- distinct(div_df, pol_group)[[1,1]]
  samp_idx <- div_df %>% 
    filter(pol_group == polx) %>% 
    group_by(across(all_of(grp_vars))) %>%
    filter(n() > samp_n)
  
  # Random sample stratified by group
  set.seed(1)
  samp_idx <- samp_idx %>% 
    do(slice_sample(., n=samp_n)) %>%
    ungroup
  
  df_samp <- div_df %>%
    filter(cell %in% samp_idx$cell)
  
  # Save
  df_samp %>% saveRDS(samp_fp)
  
  # Return
  return(df_samp)
}

get_grp_diffs <- function(df, grp_var){
  
  # Get medians and confidence intervals
  dat <- data.frame(group = df[[grp_var]], richness = df[['richness']])
  (medians <- tryCatch(
    rcompanion::groupwiseMedian(
      data = dat,
      var = "richness",
      group = "group",
      conf       = 0.95,
      R          = 1000,
      percentile = TRUE,
      bca        = FALSE,
      digits     = 3
    ),
    error = function(e) {
      rcompanion::groupwiseMedian(
        data = dat,
        var = "richness",
        group = "group",
        conf       = 0.95,
        R          = 1000,
        percentile = FALSE,
        bca        = FALSE,
        digits     = 3
      ) %>% 
        mutate(
        Conf.level = NA,
        Percentile.lower = NA,
        Percentile.upper = NA
      )
    }
  ))
  
  # Test for differences between groups
  if(length(unique(df[[grp_var]])) == 1) {
    
    print('Only one group.')
    cld <- tibble(zone = unique(df[[grp_var]]), sig_group = 'a')
    
  } else {
    
    # Check whether any groups are significantly different
    kw <- kruskal.test(df[['richness']] ~ df[[grp_var]]) 
    if(kw$p.value > 0.05) print('No significantly different groups')
    
    # Post-hoc test to compare groups
    PT <- pairwise.wilcox.test(df[['richness']], df[[grp_var]], 
                               p.adjust.method = 'fdr')
    PT1 = rcompanion::fullPTable(PT$p.value)
    cld_list <- multcompView::multcompLetters(PT1,
                  compare="<",
                  threshold=0.05,  # p-value to use as significance threshold
                  Letters=letters,
                  reversed = FALSE)
    lv <- cld_list$Letters
    cld <- tibble(sig_group = factor(lv, levels = sort(unique(lv))), 
                  zone = factor(names(lv)))
  
  }

  # Join letter indicating significantly different groups to medians and CIs
  sum1 <- select(tibble(cld), zone, sig_group) %>% 
    left_join(tibble(medians), 
              by=c(zone='group'))
  
  return(sum1)
}

iterate_get_grp_diffs <-
  function(df_samp,
           grp_var,
           var2,
           grp_diffs_fp,
           by_biome = TRUE,
           overwrite = FALSE) {
    # Get group differences for each combination of biome and ANP zone
    
    if (file.exists(grp_diffs_fp) &
        !overwrite)
      return(readRDS(grp_diffs_fp))
    
    # Nest df by groups
    if (by_biome) {
      if (is.na(var2)) {
        df_nest <- df_samp %>%
          group_by(pol_group, biome) %>%
          nest()
        
      } else if (by_biome) {
        df_nest <- df_samp %>%
          group_by(pol_group, biome, .data[[var2]]) %>%
          nest()
      }
    } else {
      if (is.na(var2)) {
        df_nest <- df_samp %>%
          group_by(pol_group) %>%
          nest()
        
      } else if (by_biome) {
        df_nest <- df_samp %>%
          group_by(pol_group, .data[[var2]]) %>%
          nest()
      }
    }
    
    # Run get_grp_diffs for each group
    medians <- df_nest %>%
      # filter(biome == 'Deserts', pol_group == 'hummingbird') %>%
      mutate(sum = purrr::map(data, get_grp_diffs, grp_var)) %>%
      select(-data) %>%
      unnest(cols = sum)
    
    # significant group factor levels
    sg_levels <- sort(as.character(unique(medians$sig_group)))
    medians <- medians %>%
      mutate(sig_group = factor(sig_group, levels = sg_levels))
    
    # Save
    medians %>% saveRDS(grp_diffs_fp)
    return(medians)
  }

plot_pollin_facet <- function(df, title = NA, var2='landcover', grp_var='anp_zone',
                              rows_var = NULL, cols_var = NULL, facet_scales = 'fixed'){
  
  # Get upper limit for X
  xmax <- max(df$Percentile.upper, na.rm = TRUE)
  if(is.null(rows_var)) rows_var = var2
  
  # Set up to shade facets by var2
  selvars <- c('pol_group', 'biome')
  if(!is.na(var2)) selvars <- c(selvars, var2)
  df_rect <- df %>% select(any_of(selvars)) %>% distinct()
  
  if(grp_var == 'anp_zone'){
    
    # var_levels <- c( 'Outside buffer', 'Buffer 10 km', 'Inside NPA')
    var_levels <- c( 'Not NPA', 'NPA')
    facet_y <- 'Land cover'
    
  } else if (grp_var == 'landcover') {
    
    var_levels <- c( 'otro', 'ag', 'veg')
    facet_y <- 'ANP zone'
    
  }
  
  # Initialize ggplot
  p <- ggplot(df)
  
  # Shade facet by ANP zone or landcover
  if(var2 != 'pol_group'){
    p <- p +
      geom_rect(data=df_rect,
                aes(xmin = -Inf, xmax = Inf, ymin = -Inf, ymax = Inf,
                    fill=.data[[var2]]), alpha=0.25, color='white') +
      scale_fill_grey(name = facet_y, start=0.05, end=0.75) 
  } 
  
  p <- p +
  
    # plot CIs
    geom_errorbar(aes(y = zone, x = Median, color=sig_group,
                      xmin = Percentile.lower, xmax = Percentile.upper),
                   width = 0.3, size  = 0.5) +
    
    # plot medians
     geom_point(aes(y = zone, x = Median, color=sig_group), 
                shape = 15, size  = 2) +
     xlab("Median richness with 95% CI") +
    scale_x_continuous(n.breaks = NULL) +
    scale_y_discrete(limits = var_levels) +
    
    # Color points by significance group
    scale_color_manual(values = c('a'='#F8766D', 
                                  'ab'='#C77CFF',
                                  'b'='#00BFC4', 
                                  'bc'='#7CAE00', 
                                  'c'='#DE8C00'), 
                       name = 'Group',
                       drop = FALSE,
                       guide = NULL) +
  
    # Facet by x=pollinator group, y = ANP zone
    # facet_grid(vars(.data[[var2]]), vars(biome)) +
    facet_grid(vars(.data[[rows_var]]), vars(.data[[cols_var]]), 
               scales = facet_scales, 
               labeller = labeller(
        biome = ~ str_replace(.x, ' ', '\n'),
        pol_group = ~ str_to_title(.x)
      )) +
    expand_limits(x=0) +
    ggtitle(title) +
    theme_minimal() +
    theme(
      axis.title.x = element_blank(),
      axis.title.y = element_blank(), 
      axis.ticks.y = element_blank(),
      axis.ticks.x = element_line('lightgray'),
        axis.text.y = element_text(size = rel(0.75)),
      # axis.text.x = element_text(margin = margin(t = .2, unit = "mm")),
      # strip.text = element_blank(),
        strip.text = element_text(size=rel(0.75),
                                  margin = margin(l = 3, b=1)),
        strip.text.y = element_text(angle = 0, hjust=0), 
      strip.background = element_blank(),
      strip.switch.pad.grid = unit(0.5, 'mm'),
      legend.position = 'none',
      plot.title = element_text(size=rel(1), face='bold', hjust=0),
      panel.spacing = unit(0.5, 'mm'),
      panel.border=element_blank(),
      panel.grid.major.y=element_blank(),
      panel.ontop = FALSE,
      panel.background = element_blank(),
        plot.background = element_rect(fill = 'white'),
        plot.caption.position = 'plot',
        plot.title.position = 'plot'
      )
  
  return(p)
  
}

bees_theme <- function() {
  theme(
    panel.grid = element_line(color = 'lightgray'),
    panel.grid.minor.x = element_line(color = 'lightgray'),
    panel.background = element_rect(fill = alpha('#A6A6A6', 0.3),
                                    color = 'white'),
    plot.background = element_rect(fill = 'white', color = NA),
    strip.text = element_text(size = rel(0.75), 
                              margin = margin(3,3,3,3)),
    strip.text.x = element_text(vjust = 0),
    # axis.text.x = element_text( hjust = c(0, 0.5, 1)),
    axis.title.x = element_text(margin = margin(t=3))
  )
}

ggsave_19cm <- function(fig_fp, p, height = 10, width = 19, units = 'cm', dpi = 600, overwrite = TRUE, ...) {
  if(file.exists(fig_fp) & !overwrite) return()
  ggsave(fig_fp, p, height = height, width = width, units = units, dpi = dpi, ...)
}
```

### Richness data frames

Convert rasters to dataframes of richness by ANP zone, biome, land cover, and pollinator group.

```{r make div_df}
# Load stacked zones
combo_ras <- terra::rast(zones_ras_fp)

rich_dir <- file.path(pred_dir, 'richness')
df_fp <- file.path(rich_dir, paste0('richness_df_', zones_code, '.rds'))
div_df <- create_richness_df(rich_dir, div_dir, zones_lu_fp, combo_ras, df_fp, 
                             zones_code = zones_code, overwrite = FALSE)
div_df <- div_df %>% 
  mutate(biome = factor(biome, levels = bioms_english_sorted),
         pol_group = factor(pol_group, 
                            levels = c('bee', 'butterfly', 'moth', 
                                       'bat', 'hummingbird', 'fly', 'wasp')))
```

#### Richness by pollinator group

```{r richness}
div_df %>%
  group_by(pol_group) %>% 
  summarise(mean = mean(richness, na.rm = TRUE),
            median = median(richness),
            min = min(richness, na.rm = TRUE), 
            max = max(richness, na.rm = TRUE)) %>% 
  arrange(desc(mean))
```

#### Richness in each ecoregion

```{r richness stats}
div_df %>%
  group_by(biome, pol_group) %>% 
  summarise(mean = mean(richness, na.rm = TRUE),
            median = median(richness),
            min = min(richness, na.rm = TRUE), 
            max = max(richness, na.rm = TRUE)) %>% 
  arrange(desc(mean))
```

#### Richness in relation to NPA

```{r richness npa}
anp_ras_fp <- str_c(tools::file_path_sans_ext(anp_terr_fp), '.tif')
r_terr <- terra::rast(anp_ras_fp)
anp_ras_fp <- str_c(tools::file_path_sans_ext(anp_zones_fp), '.tif')
r_zones <- terra::rast(anp_ras_fp)

plot(r_zones)

anp_ras_fp <- str_c(tools::file_path_sans_ext(anp_terr_fp), '.tif')
combo_ras <- c(combo_ras, ID_ANP = terra::rast(anp_ras_fp))

# div_dir <- file.path(pred_dir, 'diversity_by_zones', 'anp_ids')
anp_id_df <- create_richness_df_with_ids(rich_dir, div_dir, zones_lu_fp, combo_ras,
                                         zones_code = zones_code, overwrite = TRUE)
anp_id_df <- anp_id_df %>% 
  mutate(pol_group = factor(pol_group, 
                            levels = c('bee', 'butterfly', 'moth', 
                                       'bat', 'hummingbird', 'fly', 'wasp')))

anp_id_df %>%
  filter(pol_group == 'bee') %>% 
  group_by(pol_group, ID_ANP) %>% 
  summarise(mean = mean(richness, na.rm = TRUE),
            median = median(richness),
            min = min(richness, na.rm = TRUE), 
            max = max(richness, na.rm = TRUE)) %>% 
  arrange(desc(mean))
```

#### Sample DF

Create stratified random sample of pollinator richness values. `r samp_n` pixels are sampled from each unique combination of ecoregion, land cover type, and ANP zone. The same pixels are sampled for each pollinator group, i.e. each pollinator richness layer.

```{r sample div_df}
(n_pols <- length(levels(div_df$pol_group)))
fn_prefix <- str_c('rf', rf_vers, '_', n_pols, 'pols')

if(!is.na(samp_n)) {
  # Stratified random sample
  samp_dir <- file.path(div_dir, str_c('samp', samp_n))
  dir.create(samp_dir, recursive = TRUE, showWarnings = )
  samp_fp <- file.path(samp_dir, str_c(fn_prefix, '.rds'))
  df_samp <- sample_df(div_df, samp_n, samp_fp, 
                       grp_vars = c('anp_zone', 'biome', 'landcover'),
                       overwrite = FALSE)
} else {
  samp_dir <- file.path(div_dir, 'no_samp')
  samp_fp <- file.path(samp_dir, str_c(fn_prefix, '.rds'))
  df_samp <- div_df
}
```

### Compare richness by zones

The function `get_grp_diffs` evaluates the statistical significance of groups and returns medians, CIs, and significant groups for plotting. It will be applied to each group of pollinator, biome, and either ANP zone or land cover type.

#### Compare ANP zones

```{r compare ANP, fig.width=7, fig.height=4, cache=T}
grp_var <- 'anp_zone'
var2 <- NA
by_biome <- FALSE

# File name
var2_code <- ifelse(is.na(var2), '', str_c(str_remove_all(var2, '_.*'), '_'))
grp_var_code <- str_remove_all(grp_var, '_.*')
by_biom_code <- ifelse(by_biome, '', 'noBiome_')
grp_diffs_fp <- file.path(samp_dir, 
                          str_c('grpdiffs_', fn_prefix, '_', by_biom_code, var2_code, grp_var_code, '.rds'))
medians_anp <- iterate_get_grp_diffs(df_samp, grp_var, var2, grp_diffs_fp, 
                                     by_biome = by_biome,
                                     overwrite = FALSE)

# Plot
(
  p_anp <- medians_anp %>%
    plot_pollin_facet(
      var2 = 'pol_group',
      grp_var = grp_var,
      cols_var = 'pol_group',
      facet_scales = 'free_x'
    ) +
    facet_grid(
      cols = vars(pol_group),
      scales = 'free_x',
      labeller = labeller(
        biome = ~ str_replace(.x, ' ', '\n'),
        pol_group = ~ str_to_title(.x)
      )
    ) +
    scale_x_continuous(n.breaks= 3) +
    expand_limits(x=0) +
    bees_theme() +
    theme(plot.title = element_blank())
)

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '.png'),
                p_anp,
                height = 4,
            overwrite = TRUE)
```

#### Compare ANP zones within each biome

```{r compare ANP by biome, fig.width=7, fig.height=4, cache=T}
grp_var <- 'anp_zone'
var2 <- NA

# File name
var2_code <- ifelse(is.na(var2), '', 
                    str_c(str_remove_all(var2, '_.*'), '_'))
grp_var_code <- str_remove_all(grp_var, '_.*')
grp_diffs_fp <- file.path(samp_dir, 
                          str_c('grpdiffs_', fn_prefix, '_', 
            var2_code, grp_var_code, '.rds'))
medians_anp <- iterate_get_grp_diffs(df_samp, grp_var, var2, grp_diffs_fp, overwrite=TRUE)

# Plot
(
  p_anp <- medians_anp %>%
    plot_pollin_facet(
      'All',
      var2 = 'pol_group',
      grp_var = grp_var,
      rows_var = 'biome',
      cols_var = 'pol_group',
      facet_scales = 'free_x'
    ) +
    facet_grid(
      vars(biome),
      vars(pol_group),
      scales = 'free_x',
      labeller = labeller(
        biome = ~ str_replace(.x, ' ', '\n'),
        pol_group = ~ str_to_title(.x)
      )
    ) +
    scale_x_continuous(n.breaks= 3) +
    expand_limits(x=0) +
    labs(caption = caption) +
    bees_theme() +
    theme(plot.title = element_blank())
)

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '_polxbiome.png'),
                p_anp,
                height = 10,
            overwrite = TRUE)
```

#### Just look at bees

```{r bees compare ANP by biome, fig.width=7, fig.height=4, cache=T}
# Plot only bees
(
  p_bees <- medians_anp %>%
  filter(pol_group == 'bee') %>% 
    plot_pollin_facet(
      'Bees',
      var2 = 'pol_group',
      grp_var = grp_var
    ) +    
    facet_grid(
      vars(biome),
      vars(pol_group),
      labeller = labeller(
        biome = ~ str_replace(.x, ' ', '\n'),
        pol_group = ~ str_to_title(.x)
      )
    ) +
    scale_x_continuous(n.breaks= 3) +
    expand_limits(x=0) +
    bees_theme() +
    theme(strip.text.x = element_blank())
)

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '_bees.png'),
                p_bees,
                width = 8,
            height = 13,
            overwrite = TRUE)
```

### Compare bee subgroups

```{r bees make div_df}
# Load stacked zones
combo_ras <- terra::rast(zones_ras_fp)

rich_dir <- file.path(pred_dir, 'richness', 'bee_subgroups')
bees_df <- create_richness_df(rich_dir, div_dir, zones_lu_fp, combo_ras, overwrite = FALSE)
bees_df <- bees_df %>%
  mutate(
    pol_group = factor(
      pol_group,
      levels = c(
        'social',
        'solitary-social',
        'solitary',
        'above-ground',
        'stems',
        'cavities',
        'ground'
      )
    ),
    grouper = ifelse(
      pol_group %in% c('social', 'solitary-social', 'solitary'),
      'by sociality',
      'by nesting'
    ),
    biome = factor(biome, levels = bioms_english_sorted)
  ) 

# Stratified random sample
(n_pols <- length(levels(bees_df$pol_group)))
fn_prefix <- str_c('rf', rf_vers, '_', n_pols, 'beegroups')
samp_fp <- file.path(samp_dir, str_c(fn_prefix, '.rds'))
df_samp <- sample_df(bees_df, samp_n, samp_fp, 
                     grp_vars = c('anp_zone', 'biome', 'landcover'),
                     overwrite = FALSE)
```

```{r bees compare ANP, fig.width=7, fig.height=4, cache=T}
grp_var <- 'anp_zone'
var2 <- NA
by_biome <- FALSE

# File name
var2_code <- ifelse(is.na(var2), '', str_c(str_remove_all(var2, '_.*'), '_'))
grp_var_code <- str_remove_all(grp_var, '_.*')
by_biom_code <- ifelse(by_biome, '', 'noBiome_')
grp_diffs_fp <- file.path(samp_dir, 
                          str_c('grpdiffs_', fn_prefix, '_', by_biom_code, var2_code, grp_var_code, '.rds'))

# Get group medians and significance
medians_anp <- iterate_get_grp_diffs(df_samp, grp_var, var2, grp_diffs_fp, 
                                     by_biome = by_biome,
                                     overwrite = TRUE)
medians_anp <- medians_anp %>%
  mutate(
    grouper = ifelse(
      pol_group %in% c('social', 'solitary-social', 'solitary'),
      'by sociality',
      'by nesting'
    )
  ) 

# Plot
(p_nest <- medians_anp %>%
    filter(grouper == 'by nesting') %>% 
    plot_pollin_facet(
      title = 'Bees by nesting behavior', 
      var2 = 'pol_group',
      grp_var = grp_var,
      rows_var = 'pol_group',
      facet_scales = 'free_x'
    ) +
    facet_grid(
      rows = vars(pol_group),
      labeller = labeller(
        biome = ~ str_replace(.x, ' ', '\n'),
        pol_group = ~ str_to_title(.x)
      )
    ) +
    scale_x_continuous(n.breaks= 3) +
    expand_limits(x=0) +
    bees_theme()
)

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '_nesting.png'),
                p_nest,
                width = 8,
            height = 7,
            overwrite = TRUE)

 # Plot by sociality
(p_soc <- medians_anp %>%
    filter(grouper == 'by sociality') %>% 
    plot_pollin_facet(
      title = 'Bees by sociality', 
      var2 = 'pol_group',
      grp_var = grp_var,
      rows_var = 'pol_group',
      facet_scales = 'free_x'
    ) +
    facet_grid(
      rows = vars(pol_group),
      labeller = labeller(
        biome = ~ str_replace(.x, ' ', '\n'),
        pol_group = ~ str_to_title(.x)
      )
    ) +
    scale_x_continuous(n.breaks= 3) +
    expand_limits(x=0) +
    bees_theme()
)

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '_sociality.png'),
                p_soc,
                width = 8,
            height = 6,
            overwrite = TRUE)
```

```{r bees compare-anp, fig.width=7, fig.height=4, cache=T}
grp_var <- 'anp_zone'
grp_var_code <- 'ANPzone'
var2 <- NA
var2_code <- 'noLC'

# File name
fn <- str_c('grpdiffs_', fn_prefix, '_', var2_code, '_', grp_var_code, '.rds')
grp_diffs_fp <- file.path(samp_dir, fn)
medians_anp <- iterate_get_grp_diffs(df_samp, grp_var, var2, grp_diffs_fp, overwrite=TRUE)
medians_anp <- medians_anp %>%
  mutate(
    grouper = ifelse(
      pol_group %in% c('social', 'solitary-social', 'solitary'),
      'by sociality',
      'by nesting'
    )
  ) 

# Plot
(p_beegroups <- medians_anp %>% 
  plot_pollin_facet('Bee sub-groups by nesting behavior and sociality', 
                    var2='pol_group', grp_var=grp_var,
                    rows_var = 'biome', cols_var = 'pol_group', 
                    facet_scales = 'free_x') +
      facet_grid(vars(biome), vars(pol_group), 
               scales = 'free_x', 
               labeller = labeller(biome = ~ str_replace(.x, ' ', '\n'),
                                   pol_group = ~ str_to_title(.x))) +
    scale_x_continuous(n.breaks= 3) +
    labs(caption = caption) +
    bees_theme()
    )

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '_polxbiome.png'),
            p_beegroups, height = 10, overwrite = TRUE)

# Bees by nesting behavior
(p_bynesting <- medians_anp %>% 
  filter(pol_group %in% c('above-ground', 'cavities', 'ground', 'stems')) %>% 
  plot_pollin_facet('Bees by nesting behavior', var2='pol_group', grp_var=grp_var) +
      facet_grid(vars(pol_group), vars(biome),
               # scales = 'free_x',
               labeller = labeller(biome = ~ str_replace(.x, ' ', '\n'))) +
  bees_theme())

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '_beesnesting.png'),
            p_bynesting, height = 7, overwrite = TRUE)

# Plot only bees
(p_bysocial <- medians_anp %>% 
  filter(pol_group %in% c('social', 'solitary', 'solitary-social')) %>% 
  plot_pollin_facet('Bees by sociality', var2='pol_group', grp_var=grp_var) +
      facet_grid(vars(pol_group), vars(biome),
               # scales = 'free_x',
               labeller = labeller(biome = ~ str_replace(.x, ' ', '\n'))) +
  bees_theme())

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '_beessocial.png'),
            p_bysocial, height = 6, overwrite = TRUE)
```

### Compare all pollinators using land cover

#### Compare by landcover and NPA zone

```{r compare-landcover-x-anp, eval = FALSE}
grp_var <- 'landcover'
var2 <- 'anp_zone'
var2_code <- 'ANPzone'

# File name
fn <- str_c('grpdiffs_', fn_prefix, '_', var2_code, '_', grp_var, '.rds')
grp_diffs_fp <- file.path(samp_dir, fn)
sum2 <- iterate_get_grp_diffs(df_samp, grp_var, var2, grp_diffs_fp, overwrite=TRUE)
sum2 <- sum2 %>% mutate(biome = factor(biome, levels = bioms_english_sorted))
```

```{r plot-landcover-x-anp, fig.width=7, fig.height=8, cache=T, eval = FALSE}
# Plot just one
pol_grp <- 'bee'
(
  p_bees <- sum2 %>%
    filter(pol_group == pol_grp) %>%
    plot_pollin_facet(str_to_title(pol_grp), var2 = var2, grp_var = grp_var)  +
    facet_grid(
      vars(anp_zone),
      vars(biome),
      labeller = labeller(
        biome = ~ str_replace(.x, ' ', '\n'),
        pol_group = ~ str_to_title(.x)
      )
    ) +
    labs(caption = caption) +
    bees_theme()
)

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '_bees.png'),
            p_bees, height = 6, overwrite = TRUE)

# Plot
pol_grp <- 'hummingbird'
p_birds <- sum2 %>% 
  filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var) 

pol_grp <- 'moth'
p_moths <- sum2 %>% filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var)

pol_grp <- 'butterfly'
p_butterflies <- sum2 %>% filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var)

pol_grp <- 'fly'
p_flies <- sum2 %>% filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var)

pol_grp <- 'wasp'
p_wasps <- sum2 %>% filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var) 

pol_grp <- 'bat'
p_bats <- sum2 %>% filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var)
  
# Put the plots together
plots <- list(p_bees + theme(strip.text.x = element_text(size=rel(0.8))), 
              p_birds + theme(strip.text.x = element_text(size=rel(0.8))),
              p_butterflies, 
              p_bats, 
              p_moths + theme(axis.title.x = element_text()), 
              p_wasps,
              guide_area(), 
              p_flies + theme(axis.title.x = element_text()))

p_lc <- wrap_plots(plots, ncol=2, guides = 'collect') +
  plot_annotation(caption = caption) & 
  # scale_x_continuous(n.breaks=3,
  #                    limits = c(0, 10)) &
  theme(plot.title.position = 'plot', 
        legend.box = 'horizontal', 
        legend.position = 'right',
        axis.text.x = element_text(margin = margin(t = .2, unit = "mm"),
                                   hjust = c(0, 0.5, 1) ))

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '.png'),
            p_lc, height = 13, overwrite = TRUE)
```

#### Compare by NPA zone and landcover

```{r compare-anp-x-landcover, eval = FALSE}
grp_var <- 'anp_zone'
grp_var_code <- 'ANPzone'
var2 <- 'landcover'

# File name
fn <- str_c('grpdiffs_', fn_prefix, '_', var2, '_', grp_var_code, '.rds')
grp_diffs_fp <- file.path(samp_dir, fn)
sum2 <- iterate_get_grp_diffs(df_samp, grp_var, var2, grp_diffs_fp, overwrite=FALSE)
```

```{r plot-anp-x-landcover, fig.height=8, fig.width=7, cache=TRUE, eval = FALSE}
# Plot just one
pol_grp <- 'bee'
p_bees <- sum2 %>% 
  filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var) 

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '_bees.png'),
            p_bees, height = 6, overwrite = TRUE)

# Plot
pol_grp <- 'hummingbird'
p_birds <- sum2 %>% 
  filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var) 

pol_grp <- 'moth'
p_moths <- sum2 %>% filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var)

pol_grp <- 'butterfly'
p_butterflies <- sum2 %>% filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var)

pol_grp <- 'fly'
p_flies <- sum2 %>% filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var)

pol_grp <- 'wasp'
p_wasps <- sum2 %>% filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var) 

pol_grp <- 'bat'
p_bats <- sum2 %>% filter(pol_group == pol_grp) %>% 
  plot_pollin_facet(str_to_title(pol_grp), var2=var2, grp_var=grp_var)
  
# Put the plots together
plots <- list(p_bees + theme(strip.text.x = element_text(size=rel(0.8))), 
              p_birds + theme(strip.text.x = element_text(size=rel(0.8))),
              p_butterflies, 
              p_bats, 
              p_moths + theme(axis.title.x = element_text()), 
              p_wasps,
              guide_area(), 
              p_flies + theme(axis.title.x = element_text()))

p_anplc <- wrap_plots(plots, ncol=2, guides = 'collect') +
  plot_annotation(caption = caption) & 
  # scale_x_continuous(n.breaks=3,
  #                    limits = c(0, 10)) &
  theme(plot.title.position = 'plot', 
        legend.box = 'horizontal', 
        legend.position = 'right',
        axis.text.x = element_text(margin = margin(t = .2, unit = "mm"),
                                   hjust = c(0, 0.5, 1) ))

# Save figure
ggsave_19cm(str_c(tools::file_path_sans_ext(grp_diffs_fp), '.png'),
            p_anplc, height = 10, overwrite = TRUE)
```
